requirements.txt
anthropic>=0.40.0
streamlit>=1.40.0
python-dotenv>=1.0.0
pyyaml>=6.0
pillow>=10.0.0
pdf2image>=1.17.0
requests>=2.32.0
pandas>=2.2.0
